+++ app/src/main/java/com/sopt/now/data/repoImpl/FriendRepositoryImpl.kt
@@ -1,0 +1,13 @@
+package com.sopt.now.data.repoImpl
+
+import com.sopt.now.data.repository.FriendRepository
+import com.sopt.now.data.service.FriendService
+import com.sopt.now.presentation.friend.Friend
+
+class FriendRepositoryImpl(private val friendService: FriendService) : FriendRepository {
+
+    override suspend fun getFriends(page: Int): Result<List<Friend>> = runCatching {
+        val response = friendService.getFriends(page)
+        response.data
+    }
+}
